backend/requirements.txt
# FastAPI Core Configuration
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Database (SQLAlchemy Async Support)
SQLAlchemy[asyncio]>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
greenlet>=3.0.3

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
slowapi>=0.1.9      # Hỗ trợ rate limiting như yêu cầu

# AI, Vector Database & Third Party
openai>=1.10.0
qdrant-client>=1.7.0
payos>=0.1.3        # Thư viện tích hợp PayOS chính thức

# Utilities & Testing
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.4
